+++ Assets/Scripts/Controllers/Cloud.cs
@@ -5,13 +5,13 @@ using UnityEngine;
 public class Cloud : CBC {
   void Start() {
     // float up and down
-    this.onUpdate.AddListener(() => {
-      this.transform.SetPositionY(this.transform.position.y + Mathf.Sin(Time.time * 2) * 0.001f);
+    this.onFixedUpdate.AddListener(() => {
+      this.transform.SetPositionY(this.transform.position.y + Mathf.Sin(Time.time * 2) * 0.005f);
     });
 
     // from left to right, from -15 to 15
     this.transform.SetPositionX(-15);
-    this.onUpdate.AddListener(() => {
+    this.onFixedUpdate.AddListener(() => {
       if (this.transform.position.x < 15) {
         this.transform.SetPositionX(this.transform.position.x + 0.005f);
       } else {